[RF] move to Schema Namespace
